+++ CHANGELOG.md
@@ -1,8 +1,20 @@
 # ChangeLog #
 
-## 0.0.0 - July 20, 2015
-* Initial release.
+## 0.1.0 - September 1, 2015
+* Added tag api helpers.
+** `get_tags`, `set_tags`, `get_tag`, `set_tag`, `delete_tag` 
+
+## 0.0.4 - July 22, 2015
+* Remove debug print statement causing unauthorized clients to fail.
 
 ## 0.0.3 - July 22, 2015
 * Fixed several `is not` operators to use `!=` instead.
 * Fixed after access token expires and new creds are received during request, new token is not attached to retry request.
+
+## 0.0.0 - July 20, 2015
+* Initial release.
+
+
+
+
+